Jupyter: what are sparse models looking at?
This notebook provides visualizations of ResNet50 models pay attention
to when they classify. In other words, what are the models looking at.
This information can be used for localization, but here it is provided
merely to build our intuition.

Based on:
B. Zhou, A. Khosla, A. Lapedriza, A. Oliva, and A. Torralba. Learning
Deep Features for Discriminative Localization. CVPR'16
(arXiv:1512.04150, 2015).

https://alexisbcook.github.io/2017/global-average-pooling-layers-for-object-localization/