Fix for issue #82
The super() method of the wrong subclass was used.
In this case there were no practical implications, but we
need to move to the less error-prone syntax of Python 3.x
which does not require us to specify the subclass.

I change the super() invocations in the entire file and ran
two schedules for ResNet56 and actually got better results than
previously.  I don't think these results are related to this
change, and I cannot explain them.  Nontheless, I am committing
these new results, because I also fixed the command-line parameters
of resnet56_cifar_filter_rank_v2.yaml which had a copy & paste
error in it.